refactor lex to lexing
